[#700] Implement progress reporting in state machine CBP
