+++ Sources/ZcashLightClientKit/Block/Actions/ComputeSyncRangesAction.swift
@@ -12,7 +12,7 @@ class ComputeSyncRangesAction {
 }
 
 extension ComputeSyncRangesAction: Action {
-    func run(with context: ActionContext, didUpdate: @escaping (ActionProgress) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProgress) async -> Void) async throws -> ActionContext {
         // call internalSyncProgress and compute sync ranges and store them in context
         // if there is nothing sync just switch to finished state
