+++ Sources/ZcashLightClientKit/Block/Actions/MigrateLegacyCacheDBAction.swift
@@ -27,7 +27,7 @@ class MigrateLegacyCacheDBAction {
 }
 
 extension MigrateLegacyCacheDBAction: Action {
-    func run(with context: ActionContext, didUpdate: @escaping (ActionProgress) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProgress) async -> Void) async throws -> ActionContext {
         guard let legacyCacheDbURL = config.cacheDbURL else {
             return await updateState(context)
         }
