+++ Sources/ZcashLightClientKit/Block/Actions/SaplingParamsAction.swift
@@ -12,7 +12,7 @@ class SaplingParamsAction {
 }
 
 extension SaplingParamsAction: Action {
-    func run(with context: ActionContext, didUpdate: @escaping (ActionProgress) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProgress) async -> Void) async throws -> ActionContext {
         // Download files with sapling params.
 
         await context.update(state: .scanDownloaded)
