+++ Sources/ZcashLightClientKit/Block/CompactBlockProcessor.swift
@@ -12,81 +12,6 @@ import Combine
 
 public typealias RefreshedUTXOs = (inserted: [UnspentTransactionOutputEntity], skipped: [UnspentTransactionOutputEntity])
 
-public enum CompactBlockProgress {
-    case syncing(_ progress: BlockProgress)
-    case enhance(_ progress: EnhancementProgress)
-    case fetch
-    
-    public var progress: Float {
-        switch self {
-        case .syncing(let blockProgress):
-            return blockProgress.progress
-        case .enhance(let enhancementProgress):
-            return enhancementProgress.progress
-        default:
-            return 0
-        }
-    }
-    
-    public var progressHeight: BlockHeight? {
-        switch self {
-        case .syncing(let blockProgress):
-            return blockProgress.progressHeight
-        case .enhance(let enhancementProgress):
-            return enhancementProgress.lastFoundTransaction?.minedHeight
-        default:
-            return 0
-        }
-    }
-    
-    public var blockDate: Date? {
-        if case .enhance(let enhancementProgress) = self, let time = enhancementProgress.lastFoundTransaction?.blockTime {
-            return Date(timeIntervalSince1970: time)
-        }
-        
-        return nil
-    }
-    
-    public var targetHeight: BlockHeight? {
-        switch self {
-        case .syncing(let blockProgress):
-            return blockProgress.targetHeight
-        default:
-            return nil
-        }
-    }
-}
-
-public struct EnhancementProgress: Equatable {
-    public let totalTransactions: Int
-    public let enhancedTransactions: Int
-    public let lastFoundTransaction: ZcashTransaction.Overview?
-    public let range: CompactBlockRange
-
-    public init(totalTransactions: Int, enhancedTransactions: Int, lastFoundTransaction: ZcashTransaction.Overview?, range: CompactBlockRange) {
-        self.totalTransactions = totalTransactions
-        self.enhancedTransactions = enhancedTransactions
-        self.lastFoundTransaction = lastFoundTransaction
-        self.range = range
-    }
-    
-    public var progress: Float {
-        totalTransactions > 0 ? Float(enhancedTransactions) / Float(totalTransactions) : 0
-    }
-
-    public static var zero: EnhancementProgress {
-        EnhancementProgress(totalTransactions: 0, enhancedTransactions: 0, lastFoundTransaction: nil, range: 0...0)
-    }
-
-    public static func == (lhs: EnhancementProgress, rhs: EnhancementProgress) -> Bool {
-        return
-            lhs.totalTransactions == rhs.totalTransactions &&
-            lhs.enhancedTransactions == rhs.enhancedTransactions &&
-            lhs.lastFoundTransaction?.id == rhs.lastFoundTransaction?.id &&
-            lhs.range == rhs.range
-    }
-}
-
 /// The compact block processor is in charge of orchestrating the download and caching of compact blocks from a LightWalletEndpoint
 /// when started the processor downloads does a download - validate - scan cycle until it reaches latest height on the blockchain.
 actor CompactBlockProcessor {
