+++ Sources/ZcashLightClientKit/Block/CompactBlockProcessorNG.swift
@@ -385,8 +385,8 @@ extension CompactBlockProcessorNG {
         //        await send(event: .finished(<#T##lastScannedHeight: BlockHeight##BlockHeight#>, <#T##foundBlocks: Bool##Bool#>))
     }
 
-    private func update(progress: ActionProgress) async {
-        // handle update of the progree
+    private func update(progress: CompactBlockProgress) async {
+        await send(event: .progressUpdated(progress))
     }
 
     private func syncStopped() async {
